ZS-14 Small change in WSDL and XSD and improved fake response

+++ osgp-domain-microgrids/src/main/java/com/alliander/osgp/domain/microgrids/Measurement.java
@@ -1,23 +1,20 @@
 package com.alliander.osgp.domain.microgrids;
 
-import java.math.BigInteger;
-
 import org.joda.time.DateTime;
 
 public class Measurement extends MeasurementIdentifier {
-    private BigInteger qualifier;
+    private int qualifier;
     private DateTime time;
     private double value;
 
-    public Measurement(final int id, final String node, final BigInteger qualifier, final DateTime time,
-            final double value) {
+    public Measurement(final int id, final String node, final int qualifier, final DateTime time, final double value) {
         super(id, node);
         this.qualifier = qualifier;
         this.time = time;
         this.value = value;
     }
 
-    public BigInteger getQualifier() {
+    public int getQualifier() {
         return this.qualifier;
     }
